# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API for user to post something",
        "title": "MyGram API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Hamid Baehaqi",
            "email": "hamid1bae1@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/photo": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get photos, auth required. without query will get authenticated photos. with query will get photos from user with id in query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo"
                ],
                "summary": "get photos from user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/photo.PhotoDetailDto"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "add photo, auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo"
                ],
                "summary": "add photo",
                "parameters": [
                    {
                        "description": "photo info",
                        "name": "photo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/photo.AddPhotoInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/photo.PhotoDto"
                        }
                    }
                }
            }
        },
        "/v1/photo/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get photo by photo id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo"
                ],
                "summary": "get photo by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "photo id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/photo.PhotoDetailDto"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "update photo, auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo"
                ],
                "summary": "update photo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "photo id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "photo info",
                        "name": "photo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/photo.UpdatePhotoInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/photo.PhotoDto"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "delete photo, auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo"
                ],
                "summary": "delete photo by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "photo id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "comment.CommentDto": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "photo_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "photo.AddPhotoInput": {
            "type": "object",
            "required": [
                "caption",
                "title",
                "url"
            ],
            "properties": {
                "caption": {
                    "type": "string",
                    "example": "caption"
                },
                "title": {
                    "type": "string",
                    "example": "title"
                },
                "url": {
                    "type": "string",
                    "example": "url"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "photo.PhotoDetailDto": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/comment.CommentDto"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "photo.PhotoDto": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "photo.UpdatePhotoInput": {
            "type": "object",
            "required": [
                "caption",
                "title"
            ],
            "properties": {
                "caption": {
                    "type": "string",
                    "example": "caption"
                },
                "title": {
                    "type": "string",
                    "example": "title"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}